Colors updated, Activities moved to new package.
