+++ FlyingSaucer/build.gradle
@@ -3,9 +3,13 @@
 buildscript {
     repositories {
         jcenter()
+        maven {
+            url 'https://repos.zeroturnaround.com/nexus/content/repositories/zt-public-releases'
+        }
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:1.2.3'
+        classpath 'com.zeroturnaround.jrebel.android:jr-android-gradle:0.8.+'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
