+++ FlyingSaucer/flyingsaucer/build.gradle
@@ -1,4 +1,5 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.zeroturnaround.jrebel.android'
 
 android {
     compileSdkVersion 22
