+++ FlyingSaucer/flyingsaucer/src/main/java/com/dat/flyingsaucer/store/StoreBeersActivity.java
@@ -1,20 +1,24 @@
 package com.dat.flyingsaucer.store;
 
-import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 
 import com.dat.flyingsaucer.R;
 
 /**
  * Created by davidthacker on 6/4/15.
  */
-public class StoreBeersActivity extends FragmentActivity {
+public class StoreBeersActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.activity_store_beers);
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        toolbar.setTitle("Flying Saucer");
+        setSupportActionBar(toolbar);
     }
 }
