+++ FlyingSaucer/flyingsaucer/src/main/java/com/dat/flyingsaucer/ui/BaseActivity.java
@@ -1,0 +1,9 @@
+package com.dat.flyingsaucer.ui;
+
+import android.support.v7.app.AppCompatActivity;
+
+/**
+ * Created by davidthacker on 6/4/15.
+ */
+public class BaseActivity extends AppCompatActivity {
+}
