+++ FlyingSaucer/flyingsaucer/src/main/java/com/dat/flyingsaucer/ui/BeerListActivityFragment.java
@@ -1,4 +1,4 @@
-package com.dat.flyingsaucer;
+package com.dat.flyingsaucer.ui;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -8,6 +8,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import com.dat.flyingsaucer.R;
 import com.dat.flyingsaucer.data.dao.BeerDao;
 import com.dat.flyingsaucer.data.dao.StoreBeerItemDao;
 import com.dat.flyingsaucer.data.dao.StoreDao;
