+++ FlyingSaucer/flyingsaucer/src/main/res/layout/activity_beer_list.xml
@@ -1,7 +1,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment"
-    android:name="com.dat.flyingsaucer.BeerListActivityFragment"
+    android:name="com.dat.flyingsaucer.ui.BeerListActivityFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:layout="@layout/fragment_beer_list" />
